HackHealth-backend/requirements.txt
packaging==20.9
spacy==3.0.1
newspaper3k==0.2.8
numpy==1.20.1
pyarrow==3.0.0
pandas==1.2.2
torch_optimizer==0.1.0
tqdm
transformers==4.3.2
torch==1.7.1
simpletransformers==0.60.8
pytorch_lightning==1.1.8
requests==2.25.1
beautifulsoup4==4.9.3
datasets==1.2.1
fastapi==0.63.0
pyrouge==0.1.3
rouge_score==0.0.4
